src/tutorials/the_cathedral/requirements.txt
# Core dependencies
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dotenv==0.9.9
fonttools==4.58.1
fuzzywuzzy==0.18.0
idna==3.10
kiwisolver==1.4.8
Levenshtein==0.27.1
matplotlib==3.10.3  # For creating and displaying visualizations
networkx==3.2.1  # For graph operations and symbolic field analysis
numpy>=1.26.0,<1.29.0  # For numerical operations and image processing
packaging==25.0
pandas==2.2.3
pillow==11.2.1  # For image processing and GUI display
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
scipy==1.12.0  # For scientific computing and signal processing
seaborn==0.13.2  # For statistical visualizations
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0

# Optional dependencies
# Uncomment if using OpenAI/Anthropic APIs
# openai>=1.12.0
# anthropic>=0.8.1

# Development dependencies
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1 